# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/babies/{babyId}/media/uploads": {
      "post": {
        "operationId": "MediaController_createUpload",
        "parameters": [
          {
            "name": "babyId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "413": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/{mediaId}/complete": {
      "post": {
        "operationId": "MediaController_complete",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteMediaUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/media/{mediaId}": {
      "get": {
        "operationId": "MediaController_get",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      },
      "delete": {
        "operationId": "MediaController_abandon",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "media"
        ]
      }
    },
...
    "/api/v1/babies/{babyId}/growth/measurements": {
      "get": {
        "operationId": "GrowthController_measurements",
        "parameters": [
          {
            "name": "babyId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrowthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "growth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Baby MP API",
    "description": "Baby MP HTTP API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "birthDate"
          },
          "reason": {
            "type": "string",
            "example": "must be a valid date"
          }
        },
        "required": [
          "reason"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "AUTH_REQUIRED",
              "REFRESH_TOKEN_INVALID",
              "FORBIDDEN",
              "RESOURCE_NOT_FOUND",
              "VALIDATION_FAILED",
              "VERSION_CONFLICT",
              "IDEMPOTENCY_CONFLICT",
              "CONFLICT",
              "LAST_ADMIN_REQUIRED",
              "INVITE_INVALID",
              "INVITE_EXPIRED",
              "INVITE_REVOKED",
              "INVITE_ALREADY_USED",
              "ALREADY_A_MEMBER",
              "EXPORT_NOT_READY",
              "EXPORT_EXPIRED",
              "RATE_LIMITED",
              "UPLOAD_INCOMPLETE",
              "UPLOAD_TOO_LARGE",
              "UNSUPPORTED_MEDIA_TYPE",
              "INTERNAL_ERROR"
            ],
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "example": "提交内容有误"
          },
          "requestId": {
            "type": "string",
            "example": "req_0198f76a-cd42-7000-8000-000000000001"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetailDto"
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CreateMediaUploadDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "maxLength": 255,
            "example": "photo.jpg"
          },
          "mimeType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png"
            ]
          },
          "sizeBytes": {
            "type": "number",
            "minimum": 1,
            "maximum": 20971520
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{64}$"
          }
        },
        "required": [
          "fileName",
          "mimeType",
          "sizeBytes"
        ]
      },
      "UploadInstructionDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "PUT"
            ]
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "method",
          "url",
          "headers",
          "expiresAt"
        ]
      },
      "MediaUploadDataDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string",
            "format": "uuid"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadInstructionDto"
          }
        },
        "required": [
          "mediaId",
          "upload"
        ]
      },
      "MediaUploadResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MediaUploadDataDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "CompleteMediaUploadDto": {
        "type": "object",
        "properties": {
          "width": {
            "type": "number",
            "minimum": 1,
            "maximum": 20000
          },
          "height": {
            "type": "number",
            "minimum": 1,
            "maximum": 20000
          }
        },
        "required": [
          "width",
          "height"
        ]
      },
      "MediaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "mimeType": {
            "type": "string"
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "sizeBytes": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "uploaded",
              "ready",
              "failed",
              "deleted"
            ]
          },
          "accessUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "id",
          "mimeType",
          "width",
          "height",
          "sizeBytes",
          "status",
          "accessUrl"
        ]
      },
      "MediaResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MediaDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "GrowthPointDto": {
        "type": "object",
        "properties": {
          "recordId": {
            "type": "string",
            "format": "uuid"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          }
        },
        "required": [
          "recordId",
          "occurredAt",
          "value"
        ]
      },
      "GrowthSeriesDto": {
        "type": "object",
        "properties": {
          "metric": {
            "type": "string",
            "enum": [
              "height",
              "weight"
            ]
          },
          "unit": {
            "type": "string",
            "enum": [
              "cm",
              "kg"
            ]
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrowthPointDto"
            }
          }
        },
        "required": [
          "metric",
          "unit",
          "points"
        ]
      },
      "GrowthResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GrowthSeriesDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}